backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
langchain>=0.3.27
langchain-groq>=0.3.8
langgraph>=0.6.10
groq>=0.32.0
pydantic>=2.12.0
typing-extensions>=4.15.0
python-dotenv>=1.0.0
